requirements.txt
annexremote==1.5.0
appdirs==1.4.4
astor==0.8.1
bids-validator==1.8.9
boto==2.49.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
cycler==0.11.0
datalad==0.15.4
Deprecated==1.2.13
distro==1.6.0
docopt==0.6.2
fasteners==0.17.2
fonttools==4.28.5
formulaic==0.2.4
future==0.18.2
humanize==3.13.1
idna==3.3
importlib-metadata==4.10.1
interface-meta==1.2.4
iso8601==1.0.2
keyring==23.5.0
keyrings.alt==4.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
msgpack==1.0.3
nibabel==3.2.1
num2words==0.5.10
numpy==1.22.1
packaging==21.3
pandas==1.4.0
patool==1.12
Pillow==9.0.0
pybids==0.14.0
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-gitlab==3.1.0
pytz==2021.3
requests==2.27.1
requests-toolbelt==0.9.1
scipy==1.7.3
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.3.24
tqdm==4.62.3
urllib3==1.26.8
Whoosh==2.7.4
wrapt==1.13.3
zipp==3.7.0